Do not guess type, don't track type
As noted by @jhpoelen in #9, it's probably best to drop `type` for now.  Guessing by mime type alone is really not desirable, and we're not using this information in any way in the core implementation.

type can probably be tracked in richer form in accompanying metadata/prov record for the content.

This also lets us drop a dependency (mime).

Also fix a few easy naming things. (query argument should be consistently called `uri` across query, query_local, and query_remote; dublincore format renamed to indicate it is just a formatter for hash-archive.org's return type.