Have store() use seperate registry for local content
This seperate registry follows the BagIt format, which may be a mistake.  Perhaps it would be better to follow the same registry format as before, but just use separate files (and relative file paths)?

The motivation for BagIt here is that it's a well-defined standard that matches the pattern of publishing a store pretty closely.  (Of course, since content path is just derived from the sha-256 hash, the manifest is pretty redundant.  but BagIt does have a well defined convention for the 'root' of a local/relative path, which is why I'm using it in the first place....