pin with option to skip verification
This permits pin() to act like a fast cache (avoiding the download & hash time required by hash-archive.org registration) if the user is happy to trust that the content isn't changing at the URL.  This uses query to find any registered hash for the content.

Relative to verify=TRUE, this will allow on-par performance with `pins::pins()`.  Relative to `pins`, the use of query will be slightly slower, but has the potential if unlikely advantage to find a newly registered URL for the same content, if the requested URL is no longer available but a new URL has been registered.