Divide query() into history() and sources()
query() was a vague verb that we used for both "querying on content id" and "querying on url".  query() still exists as before (though may later be deprecated), but I now expose these different types of queries as different functions:

history() takes a URL and returns content seen at that URL.

sources() takes a content uri and returns sources (file paths and URLs) for that content.

These terms match the endpoint names used in hash-archive.org.  This PR also re-factors things a bit so that the component methods for local registeries and for hash-archive.org are defined first in terms of these lower-level verbs sources() and history().

I think this should be make for a more easily-extensible model.  For instance, a pure content store (local or remote, like deeplinker.bio) can provide a sources() response given a URI, but cannot necessarily provide history() by itself (of course an associated registry could provide that).

closes #14